+++ Assets/UniGLTF/Runtime/UniGLTF/IO/MeshExportSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26b97430e1698e94e9a27ab322f3eee5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
