+++ Assets/UniGLTF/Runtime/UniGLTF/IO/MeshExporterDivided.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92bc82be5ead84a429346379590f6d7d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
